Protect hex input from invalid chars
Errors are thrown otherwise, and even though they're
silent, it's just not a nice thing to have happening.

Ideally we'd return false on invalid chars in keydown;
that method doesn't have a visual backspacing effect.

But then we'd have to deal with heaps of keycodes, and
make sure we don't affect the usual stuff like all the
functions keys, backspace, etc. And they can vary among
among different keyboards. It's quite simply not worth
the pain.